Text drawing - a fix for the case with provided width only
If width only was provided - it was possible that the text
height will be lost, and for the case of 'Center' alignment
it means that it would be centered around the 1st line also -
just 2nd part of the text will be visible.

Fixed by additional check to prevent such situation.